modules: fix issues with dropped answers - resolve()
Well... practically it still can't happen that an internal request
gets its answer dropped, but ATM my understanding of the API is that
it is allowed to happen already, and the crashes during tests were
bothering me (simulating drops).

This may become more relevant in future, e.g. if we allow dropping
as a policy action; policy authors may not care about the request being
internal.