cache: don't change kr_layer_t::state to _DONE
I suspect this might've caused real-life issues in some edge case,
making a cached CNAME not being followed.  But in any case,
cache is not a good place to set KR_STATE_DONE, as it only creates
a (pseudo-)packed for immediate CONSUME phase; it should be
iterator who sets it in that later phase when picking those records.

I added unnecessary simplifications to code, too (all those returns).