selection:  balance the choice of IPv4/IPv6 servers when RTT cache is clear
On clear RTT and in the presence of miscunfigured IP protocol version (e.g. net.ipv6 enabled
but IPv6 doesn't work), resolver might waste a lot of time and packets trying to query
nameservers using the broken protocol.

We choose the protocol first and if there is some candidate server with unknown RTT and using
the chosen protocol, we swap it to the front and it is used for the transport.

If there is none, tough luck, we still want to try all the candidates before going back trying
some again.