policy.DEBUG_*: tweak verbosity
One issue I saw in verbose(true) mode is that using DEBUG_IF with
a false condition would turn off verbosity for those requests,
and that seemed surprising.
Also, if we know in advance that we want to print the verbose lines,
let them be printed directly instead of using the complex mechanism.