fixup! lib/utils: define kr_require(), kr_assume() macros
Reduce the amount of "inlined" code and simplify.

Details: this by itself reduces code size by kilobytes and the impact
will probably be larger once we use the functions more.  It might also
perform better.  The point is to inline just the parts that are likely
to happen and put everything else into shared __cold__ code,
as generally it's advantageous to optimize cold parts for (code) size.
(the KR_COLD also nicely replaces __builtin_expect() calls)