WIP: daemon/worker: weak pointer logic for tasks
Replaces the reference-counting logic for `struct qr_task` with
weak-pointer-like logic (in `lib/weakptr`). Pointers to `struct qr_task`
outside of `daemon/worker` are replaced with a unique task identifier
(`qr_task_weakptr_t`), which may be passed to worker functions. If a
weak pointer is no longer valid, operations are (usually silently)
skipped, because it is assumed that a task that has been freed has
already been finished, and any pending operations on it are no longer
needed.

(This is a work in progress and still has some bugs that need to be
resolved)