daemon: protolayer timeout and stream fixes
Incoming TCP sessions (from clients) were not marked as `connected`,
causing the wrong timeout function to be called. This may have resulted
in assertion failures in some cases when using TCP.

Some error codes were not wrapped in `kr_error` - this probably did not
cause any real issues currently, but for the sake of consistency this is
now fixed.

Stream wire buffers were not being processed in a loop, which could
potentially cause input data to be discarded or processed with a delay.
This is now fixed.