daemon/session2: rename http flag to custom_emalf_handling
The session2 API as well as everything that uses it should be as
independent of the underlying protocols as possible, providing
functionality through firing different events. There was a left-over
`struct session2::http` flag, which allowed the DoH layer to return a
Bad Request reply. This has been renamed so that it does not mention
HTTP apart from the example in the doc comments, in case another
protocol needs to use this functionality at some point.