manager/poetry.lock: remove
Lockfiles are generally encouraged to be committed into VCS to ensure
reproducible builds on all machines. Our situation is slightly
different - we are more interested in supporting a wide range of setups
on a wide range of systems.

Our `poetry.lock` also contained some outdated library versions that had
CVEs reported on them, and GitHub did not like that, so this
incidentally also solves that problem for us. Since the lock file was
only used during development, this will not affect end users in any way
(runtime dependencies are managed by each distro's package manager).