modules/hints,lib/rules: rework TTL defaults
It was a bit wide, with 5s and 3h, and distinction between
these two "groups" of rules was a bit random wrt. TTL choice.
Now: 5m for user's rules and 1h for RFC-default rules.

I found it relatively hard to choose defaults, but at least for
user-supplied rules it's trivial to configure a different default.