daemon/proxyv2: move PROXY protocol into its own layer
Previously, PROXYv2 handling was partially implemented in the `io.c`
unit in the `_TCP` and `_UDP` protocol layers, which technically made
very little sense. This commit moves this handling into separate
`_PROXYV2_DGRAM` and `_PROXYV2_STREAM` protocol layers, basically
encapsulating the handling of proxies in the `proxyv2.c` unit.

This commit also removes support for `PROTOLAYER_PAYLOAD_IOVEC`-type
buffers from the PROXYv2 layer, as it was unused in this context.