resolver,api: set AA bit if all RRsets have (only) rank KR_RANK_AUTH.
By using logic similar to that for the AD bit, AA bit in responses is
set in resolver's answer_finalize if all RRsets are non-secure (so no
AD bit) and have rank KR_RANK_AUTH.
KR_RANK_AUTH is safe enough because in current code it is never set
alone.
If a RRset has only KR_RANK_AUTH, it indicates that an authoritative
response is wanted.

Fixes #913