ci: update to new IMAGE_TAG
For 6.x we started using this TAG in commit a1aa3a51a2ede
but there have been other IMAGE_TAG bumps in the meantime.

Here we'll need it for the deckard update in the next commit,
which is needed due to validator tweaks.
And to fix CI, let's do the tiny bump in the lua bindings
(no practical impact whatsoever due to struct alignment).