datamodel: also drop /dnssec/keep-removed
This is basically an amendment of 45d95e3b.

I consider RFC 5011 pretty exotic to use, and even for the root zone
there are typically better options.  So exposing tiny tweaks
in the 5011 algoritm seems to bring more confusion than value.
(basically same situation as for the other /dnssec/* removals)