treewide: replace port 5353 by 5335
My main aim was not to have it in documentation,
but I ended up removing it from everywhere except ci/respdiff
(which I didn't feel like messing with right now).

I see people using 5353 for DNS, and it's problematic,
as the port is reserved for mDNS which is quite common,
and the packets even have similar format which makes it only worse.