doc/dev: reinstante the config-debugging section
It apparently got forgotten when splitting the doc/ into doc/{user,dev}/
And as we now removed this from YAML (and doc/user) in 9fbacef1709,
these options completely disappeared from our 6.x docs.
So this PR puts it back into doc/dev on the place where it's in 5.x docs.